# Conflicts:
#	trace-sdk/src/androidTest/java/io/bitrise/trace/data/collector/cpu/ApplicationCpuUsageDataCollectorInstrumentedTest.java
#	trace-sdk/src/androidTest/java/io/bitrise/trace/data/collector/cpu/SystemCpuUsageDataCollectorInstrumentedTest.java
#	trace-sdk/src/androidTest/java/io/bitrise/trace/data/collector/device/DeviceCarrierDataCollectorInstrumentedTest.java
#	trace-sdk/src/androidTest/java/io/bitrise/trace/data/collector/device/DeviceIdDataCollectorInstrumentedTest.java
#	trace-sdk/src/androidTest/java/io/bitrise/trace/data/collector/device/DeviceLocaleDataCollectorInstrumentedTest.java
#	trace-sdk/src/androidTest/java/io/bitrise/trace/data/collector/device/DeviceModelDataCollectorInstrumentedTest.java
#	trace-sdk/src/androidTest/java/io/bitrise/trace/data/collector/device/DeviceNetworkTypeDataCollectorInstrumentedTest.java
#	trace-sdk/src/androidTest/java/io/bitrise/trace/data/collector/device/DeviceOsVersionDataCollectorInstrumentedTest.java
#	trace-sdk/src/androidTest/java/io/bitrise/trace/data/collector/memory/ApplicationUsedMemoryDataCollectorInstrumentedTest.java
#	trace-sdk/src/androidTest/java/io/bitrise/trace/data/collector/memory/SystemMemoryDataCollectorInstrumentedTest.java
#	trace-sdk/src/androidTest/java/io/bitrise/trace/data/collector/view/ApplicationStartUpDataListenerInstrumentedTest.java
#	trace-sdk/src/androidTest/java/io/bitrise/trace/data/collector/view/FragmentStateDataListenerInstrumentedTest.java
#	trace-sdk/src/androidTest/java/io/bitrise/trace/network/MetricSenderIntegrationTest.java
#	trace-sdk/src/main/java/io/bitrise/trace/configuration/ConfigurationManager.java
#	trace-sdk/src/main/java/io/bitrise/trace/data/collector/TraceLogCollector.java
#	trace-sdk/src/main/java/io/bitrise/trace/data/collector/cpu/SystemCpuUsageDataCollector.java
#	trace-sdk/src/main/java/io/bitrise/trace/data/collector/device/DeviceCarrierDataCollector.java
#	trace-sdk/src/main/java/io/bitrise/trace/data/collector/device/DeviceNetworkTypeDataCollector.java
#	trace-sdk/src/main/java/io/bitrise/trace/data/collector/memory/MemoryDataCollector.java
#	trace-sdk/src/main/java/io/bitrise/trace/data/collector/network/NetworkDataListener.java
#	trace-sdk/src/main/java/io/bitrise/trace/data/collector/network/TraceNetworkListener.java
#	trace-sdk/src/main/java/io/bitrise/trace/data/collector/view/ApplicationStartUpDataListener.java
#	trace-sdk/src/main/java/io/bitrise/trace/data/collector/view/FragmentStateDataListener.java
#	trace-sdk/src/main/java/io/bitrise/trace/data/dto/FragmentData.java
#	trace-sdk/src/main/java/io/bitrise/trace/data/dto/FragmentDataStateEntry.java
#	trace-sdk/src/test/java/io/bitrise/trace/data/collector/cpu/SystemCpuUsageDataCollectorTest.java
#	trace-sdk/src/test/java/io/bitrise/trace/data/collector/device/DeviceRootedDataCollectorTest.java
#	trace-sdk/src/test/java/io/bitrise/trace/data/collector/network/okhttp/OkHttpDataListenerTest.java
#	trace-sdk/src/test/java/io/bitrise/trace/data/collector/network/urlconnection/TraceHttpsURLConnectionTest.java
#	trace-sdk/src/test/java/io/bitrise/trace/data/collector/network/urlconnection/TraceOkHttpClientTest.java
#	trace-sdk/src/test/java/io/bitrise/trace/data/collector/network/urlconnection/TraceURLStreamHandlerTest.java
#	trace-sdk/src/test/java/io/bitrise/trace/data/collector/view/FragmentStateDataListenerTest.java

+++ trace-sdk/src/test/java/io/bitrise/trace/data/management/formatter/ResourceDataFormatterTest.java
@@ -1,84 +1,88 @@
 package io.bitrise.trace.data.management.formatter;
 
-import org.junit.Test;
-import org.mockito.Mockito;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 
 import io.bitrise.trace.data.dto.Data;
 import io.bitrise.trace.data.dto.FormattedData;
 import io.bitrise.trace.data.resource.ResourceLabel;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
+import org.junit.Test;
+import org.mockito.Mockito;
 
 /**
  * Tests for {@link ResourceDataFormatter}.
  */
-public class ResourceDataFormatterTest extends BaseDataFormatterTest{
+public class ResourceDataFormatterTest extends BaseDataFormatterTest {
 
-    final Data inputData = Mockito.mock(Data.class, Mockito.CALLS_REAL_METHODS);
-    final ResourceLabel dummyResourceLabel = ResourceLabel.DEVICE_ROOTED;
-    final String dummyResourceEntityLabel = "device.rooted";
+  final Data inputData = Mockito.mock(Data.class, Mockito.CALLS_REAL_METHODS);
+  final ResourceLabel dummyResourceLabel = ResourceLabel.DEVICE_ROOTED;
+  final String dummyResourceEntityLabel = "device.rooted";
 
-    @Test
-    public void formatResource_string() {
-        inputData.setContent("an exciting string resource");
+  @Test
+  public void formatResource_string() {
+    inputData.setContent("an exciting string resource");
 
-        final FormattedData[] outputArray = ResourceDataFormatter.formatResource(inputData, dummyResourceLabel);
-        assertEquals(1, outputArray.length);
-        final FormattedData outputData = outputArray[0];
+    final FormattedData[] outputArray =
+        ResourceDataFormatter.formatResource(inputData, dummyResourceLabel);
+    assertEquals(1, outputArray.length);
+    final FormattedData outputData = outputArray[0];
 
-        assertNotNull(outputData);
-        assertNotNull(outputData.getResourceEntity());
-        assertEquals(dummyResourceEntityLabel, outputData.getResourceEntity().getLabel());
-        assertEquals("an exciting string resource", outputData.getResourceEntity().getValue());
-    }
+    assertNotNull(outputData);
+    assertNotNull(outputData.getResourceEntity());
+    assertEquals(dummyResourceEntityLabel, outputData.getResourceEntity().getLabel());
+    assertEquals("an exciting string resource", outputData.getResourceEntity().getValue());
+  }
 
-    @Test
-    public void formatResource_integer() {
-        inputData.setContent((int)21);
+  @Test
+  public void formatResource_integer() {
+    inputData.setContent(21);
 
-        final FormattedData[] outputArray = ResourceDataFormatter.formatResource(inputData, dummyResourceLabel);
-        assertEquals(1, outputArray.length);
-        final FormattedData outputData = outputArray[0];
+    final FormattedData[] outputArray =
+        ResourceDataFormatter.formatResource(inputData, dummyResourceLabel);
+    assertEquals(1, outputArray.length);
+    final FormattedData outputData = outputArray[0];
 
-        assertNotNull(outputData);
-        assertNotNull(outputData.getResourceEntity());
-        assertEquals(dummyResourceEntityLabel, outputData.getResourceEntity().getLabel());
-        assertEquals("21", outputData.getResourceEntity().getValue());
-    }
+    assertNotNull(outputData);
+    assertNotNull(outputData.getResourceEntity());
+    assertEquals(dummyResourceEntityLabel, outputData.getResourceEntity().getLabel());
+    assertEquals("21", outputData.getResourceEntity().getValue());
+  }
 
-    @Test
-    public void formatResource_float() {
-        inputData.setContent((float)3.145);
+  @Test
+  public void formatResource_float() {
+    inputData.setContent((float) 3.145);
 
-        final FormattedData[] outputArray = ResourceDataFormatter.formatResource(inputData, dummyResourceLabel);
-        assertEquals(1, outputArray.length);
-        final FormattedData outputData = outputArray[0];
+    final FormattedData[] outputArray =
+        ResourceDataFormatter.formatResource(inputData, dummyResourceLabel);
+    assertEquals(1, outputArray.length);
+    final FormattedData outputData = outputArray[0];
 
-        assertNotNull(outputData);
-        assertNotNull(outputData.getResourceEntity());
-        assertEquals(dummyResourceEntityLabel, outputData.getResourceEntity().getLabel());
-        assertEquals("3.145", outputData.getResourceEntity().getValue());
-    }
+    assertNotNull(outputData);
+    assertNotNull(outputData.getResourceEntity());
+    assertEquals(dummyResourceEntityLabel, outputData.getResourceEntity().getLabel());
+    assertEquals("3.145", outputData.getResourceEntity().getValue());
+  }
 
-    @Test
-    public void formatResource_boolean() {
-        inputData.setContent(true);
+  @Test
+  public void formatResource_boolean() {
+    inputData.setContent(true);
 
-        final FormattedData[] outputArray = ResourceDataFormatter.formatResource(inputData, dummyResourceLabel);
-        assertEquals(1, outputArray.length);
-        final FormattedData outputData = outputArray[0];
+    final FormattedData[] outputArray =
+        ResourceDataFormatter.formatResource(inputData, dummyResourceLabel);
+    assertEquals(1, outputArray.length);
+    final FormattedData outputData = outputArray[0];
 
-        assertNotNull(outputData);
-        assertNotNull(outputData.getResourceEntity());
-        assertEquals(dummyResourceEntityLabel, outputData.getResourceEntity().getLabel());
-        assertEquals("true", outputData.getResourceEntity().getValue());
-    }
+    assertNotNull(outputData);
+    assertNotNull(outputData.getResourceEntity());
+    assertEquals(dummyResourceEntityLabel, outputData.getResourceEntity().getLabel());
+    assertEquals("true", outputData.getResourceEntity().getValue());
+  }
 
-    @Test
-    public void formatResource_nullContent() {
-        inputData.setContent(null);
-        final FormattedData[] outputArray = ResourceDataFormatter.formatResource(inputData, dummyResourceLabel);
-        assertEquals(0, outputArray.length);
-    }
+  @Test
+  public void formatResource_nullContent() {
+    inputData.setContent(null);
+    final FormattedData[] outputArray =
+        ResourceDataFormatter.formatResource(inputData, dummyResourceLabel);
+    assertEquals(0, outputArray.length);
+  }
 }
